button story refactor
